requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
transformers==4.44.2
torch==2.4.1
sentence-transformers==3.1.1
accelerate==0.34.2
httpx==0.27.2
python-dotenv==1.0.1
google